+++ EleFitsData/EleFitsData/DataUtils.h
@@ -203,6 +203,15 @@ constexpr decltype(auto) applyImpl(TTuple&& tuple, TFunc&& func, std::index_sequ
 }
 
 /**
+ * @brief Apply a variadic function to elements pointed by a tuple of iterators
+ * and increment the iterators.
+ */
+template <typename TTuple, typename TFunc, std::size_t... Is>
+constexpr decltype(auto) generateImpl(TTuple&& tuple, TFunc&& func, std::index_sequence<Is...>) {
+  return func(*std::get<Is>(tuple)++...);
+}
+
+/**
  * @brief Apply a function which returns void to each element of a tuple.
  */
 template <typename TTuple, typename TFunc, std::size_t... Is>
@@ -271,6 +280,18 @@ constexpr decltype(auto) tupleApply(TTuple&& tuple, TFunc&& func) {
       std::forward<TTuple>(tuple),
       std::forward<TFunc>(func),
       Internal::tupleIndexSequence<TTuple>());
+}
+
+/**
+ * @brief Apply a variadic function to elements pointed by a tuple of iterators
+ * and increment the iterators.
+ */
+template <typename TIteratorTuple, typename TFunc>
+constexpr decltype(auto) tupleGenerate(TIteratorTuple&& tuple, TFunc&& func) { // FIXME naming
+  return Internal::generateImpl(
+      std::forward<TIteratorTuple>(tuple),
+      std::forward<TFunc>(func),
+      Internal::tupleIndexSequence<TIteratorTuple>());
 }
 
 /**